correction in ex. 4

+++ src/main/java/homeworks/tinasargsyan/hw2/Exercise4.java
@@ -11,9 +11,9 @@ public class Exercise4 {
         Scanner sc = new Scanner(System.in);
         System.out.println("Please enter the number");
         int n = sc.nextInt();
-        double sum=0;
-        for (double i=1;i<=n;i++){
-            sum+=(1/i);
+        double sum = 0;
+        for (int i=1;i<=n;i++){
+            sum+=(1.0/i);
         }
         System.out.println("The sum of the following series " + sum);
     }
